updated pages to follow other route -> page -> component formating.
